# nsc-practicum-winter2018-api/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "API for T-Mobile/UST Onboarding tracker",
    "version" : "1.0.5",
    "title" : "Onboarding Tool",
    "contact" : {
      "email" : "jason.panzer@seattlecolleges.edu"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths" : {
    "/user" : {
      "post" : {
        "summary" : "Creates a new user",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "user",
          "description" : "The user to create.",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/User"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Success! User record has been created."
          },
          "400" : {
            "description" : "Bad request."
          },
          "401" : {
            "description" : "Authorization information is missing or invalid."
          },
          "403" : {
            "description" : "Operation not authorized."
          },
          "500" : {
            "description" : "Internal server error."
          },
          "501" : {
            "description" : "Service not yet implemented."
          }
        }
      }
    },
    "/user/{id}" : {
      "get" : {
        "summary" : "Retrieve a user by id.",
        "description" : "Returns a user by id. If the user is of type \"employee,\" The user object will contain the id of the workflow a user has been assigned, as well as an array of task ids representing the tasks a user has completed. If the user is of type \"manager,\" the object will contain an an associative array where the key is the employee ID and the value if the employee's name (first and last) for the employees they manage",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "description" : "The id of the user to retrieve.",
          "required" : true,
          "type" : "string"
        } ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Success! User has been found.",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          },
          "400" : {
            "description" : "Bad request. User not found."
          },
          "401" : {
            "description" : "Authorization information is missing or invalid."
          },
          "403" : {
            "description" : "Operation not authorized."
          },
          "500" : {
            "description" : "Internal server error."
          },
          "501" : {
            "description" : "Service not yet implemented."
          }
        }
      },
      "put" : {
        "summary" : "Updates an existing user.",
        "description" : "Updates the user record associated with the specified id.",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "description" : "id of the user object to update.",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "user",
          "required" : true,
          "schema" : {
            "type" : "object",
            "properties" : {
              "id" : {
                "type" : "string",
                "example" : "72AD9DBC60AE485782D43A1AE09279A"
              },
              "firstName" : {
                "type" : "string",
                "example" : "Rick"
              },
              "lastName" : {
                "type" : "string",
                "example" : "Sanchez"
              },
              "email" : {
                "type" : "string",
                "example" : "rsanchez@example.org"
              },
              "phone" : {
                "type" : "string",
                "example" : "999-989-9229"
              },
              "type" : {
                "type" : "string",
                "example" : "manager"
              },
              "employees" : {
                "type" : "object",
                "additionalProperties" : true,
                "example" : {
                  "72AD9DBC60AE485782D43A1AE09279A" : "Morty Smith",
                  "72AD9DBC60AE485782D43A1A309279A" : "Jesse"
                }
              }
            }
          }
        } ],
        "consumes" : [ "application/json" ],
        "responses" : {
          "201" : {
            "description" : "Success! User record has been updated."
          },
          "400" : {
            "description" : "Bad request. User not found."
          },
          "401" : {
            "description" : "Authorization information is missing or invalid."
          },
          "403" : {
            "description" : "Operation not authorized."
          },
          "500" : {
            "description" : "Internal server error."
          },
          "501" : {
            "description" : "Service not yet implemented."
          }
        }
      },
      "delete" : {
        "summary" : "Deletes an existing user.",
        "description" : "Deletes an existing user as denoted by the path-specified id.",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "description" : "id of user to delete.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success! User record has been removed."
          },
          "400" : {
            "description" : "Bad request. User not found."
          },
          "401" : {
            "description" : "Authorization information is missing or invalid."
          },
          "403" : {
            "description" : "Operation not authorized."
          },
          "500" : {
            "description" : "Internal server error."
          },
          "501" : {
            "description" : "Service not yet implemented."
          }
        }
      }
    },
...
    "/user/{id}/employees" : {
      "post" : {
        "summary" : "Assigns an employee to a manager.",
        "parameters" : [ {
          "in" : "path",
          "type" : "string",
          "name" : "id",
          "description" : "id of the manager.",
          "required" : true
        }, {
          "in" : "body",
          "name" : "user",
          "description" : "A new employee",
          "required" : true,
          "schema" : {
            "type" : "object",
            "required" : [ "firstName", "lastName", "email", "phone" ],
            "properties" : {
              "firstName" : {
                "type" : "string",
                "example" : "Jesse"
              },
              "lastName" : {
                "type" : "string",
                "example" : "Swarthout"
              },
              "email" : {
                "type" : "string",
                "example" : "jetti@example.org"
              },
              "phone" : {
                "type" : "string",
                "example" : "999-989-9229"
              },
              "hireDate" : {
                "type" : "string",
                "example" : "1997-07-16"
              },
              "workflow" : {
                "type" : "string",
                "example" : "72AD9DBC60AE485782D43A1AE09279A"
              }
            }
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Success! Employee has been assigned to manager."
          },
          "400" : {
            "description" : "Bad request. User not found."
          },
          "401" : {
            "description" : "Authorization information is missing or invalid."
          },
          "403" : {
            "description" : "Operation not authorized."
          },
          "500" : {
            "description" : "Internal server error."
          },
          "501" : {
            "description" : "Service not yet implemented."
          }
        }
      },
      "delete" : {
        "summary" : "Unnassigns an employee from a manager.",
        "consumes" : [ "text/plain" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "description" : "id of user.",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "task",
          "description" : "id of the user.",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "81484e47aa46445ab8ab73f01be9dcc1"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success! User record has been removed."
          },
          "400" : {
            "description" : "Bad request. User not found."
          },
          "401" : {
            "description" : "Authorization information is missing or invalid."
          },
          "403" : {
            "description" : "Operation not authorized."
          },
          "500" : {
            "description" : "Internal server error."
          },
          "501" : {
            "description" : "Service not yet implemented."
          }
        }
      }
    },
...
  },
  "definitions" : {
    "User" : {
      "description" : "A User (spoiler alert) represents a human, or more pointedly, the users that will interact with the UTC onboarding system. The User object stores basic identity deets for both employee and manager users.",
      "type" : "object",
      "required" : [ "firstName", "lastName", "email", "phone", "type" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "example" : "72AD9DBC60AE485782D43A1AE09279A4",
          "description" : "Uniquely identifies a User object."
        },
        "firstName" : {
          "type" : "string",
          "example" : "Rick",
          "description" : "User's first name."
        },
        "lastName" : {
          "type" : "string",
          "example" : "Sanchez",
          "description" : "User's last name."
        },
        "email" : {
          "type" : "string",
          "format" : "email",
          "example" : "rick.sanchez@example.org",
          "description" : "User's email address."
        },
        "phone" : {
          "type" : "string",
          "example" : "989-999-2229",
          "description" : "User's phone number."
        },
        "type" : {
          "type" : "string",
          "enum" : [ "employee", "manager" ],
          "description" : "A User's \"type\" can be one of the following - employee or manager.",
          "example" : "manager"
        },
        "startDate" : {
          "type" : "string",
          "description" : "The date of employment for a new hire.",
          "example" : "2018-02-01"
        },
        "workflow" : {
          "type" : "string",
          "description" : "The id of the workflow the user has been assigned.",
          "example" : "72AD9DBC60AE485782D43A1AE09279A4"
        },
        "tasks" : {
          "type" : "array",
          "description" : "An array of task IDs indicating which tasks the user has been completed.",
          "items" : {
            "type" : "string"
          },
          "example" : [ "ECCD3A6ED4C54D2DA28C9CDD28F6417E", "ECCD3A6ED4C54D2DA28C9CDD28F6417E", "ECCD3A6ED4C54D2DA28C9CDD28F6417E" ]
        },
        "employees" : {
          "type" : "object",
          "additionalProperties" : true,
          "example" : {
            "72AD9DBC60AE485782D43A1AE09279A" : "Morty Smith",
            "72AD9DBC60AE485782D43A1A309279A" : "Toddy"
          }
        }
      }
    },
...
  },
  "host" : "virtserver.swaggerhub.com",
  "schemes" : [ "https" ],
  "basePath" : "/kari_bullard/Cloud-Practicum/1.0.5"
}